chore(deps): batch jsonschema, tree-sitter and action-gh-release (#155)
* chore(deps): bump softprops/action-gh-release from 3.0.2 to 3.0.3

Bumps [softprops/action-gh-release](https://github.com/softprops/action-gh-release) from 3.0.2 to 3.0.3.
- [Release notes](https://github.com/softprops/action-gh-release/releases)
- [Changelog](https://github.com/softprops/action-gh-release/blob/master/CHANGELOG.md)
- [Commits](https://github.com/softprops/action-gh-release/compare/3d0d9888cb7fd7b750713d6e236d1fcb99157228...efb35369e0ad2afab669f228072c1b0d510eae64)

---
updated-dependencies:
- dependency-name: softprops/action-gh-release
  dependency-version: 3.0.3
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* chore(deps): bump tree-sitter from 0.26.13 to 0.27.0

Bumps [tree-sitter](https://github.com/tree-sitter/tree-sitter) from 0.26.13 to 0.27.0.
- [Release notes](https://github.com/tree-sitter/tree-sitter/releases)
- [Commits](https://github.com/tree-sitter/tree-sitter/compare/v0.26.13...v0.27.0)

---
updated-dependencies:
- dependency-name: tree-sitter
  dependency-version: 0.27.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* chore(deps): bump jsonschema from 0.51.0 to 0.52.1

Bumps [jsonschema](https://github.com/Stranger6667/jsonschema) from 0.51.0 to 0.52.1.
- [Release notes](https://github.com/Stranger6667/jsonschema/releases)
- [Changelog](https://github.com/Stranger6667/jsonschema/blob/master/CHANGELOG.md)
- [Commits](https://github.com/Stranger6667/jsonschema/compare/cli-v0.51.0...ruby-v0.52.1)

---
updated-dependencies:
- dependency-name: jsonschema
  dependency-version: 0.52.1
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Drop casts that tree-sitter 0.27 made redundant

`Node::child_count` returns `u32` in 0.27 where it returned `usize` before, so
the four `index as u32` casts feeding `child` and `field_name_for_child` are now
`u32` to `u32` and clippy denies them. The bump builds and every test passes
either way; this is what turns the workspace clippy gate green again.

Nothing about which children are selected changes.


---------

+++ crates/agent-guard/src/shell.rs
@@ -3087,10 +3087,10 @@ fn tee_output_targets(node: tree_sitter::Node<'_>, source: &[u8]) -> Vec<String>
 fn output_redirect_targets(redirected: tree_sitter::Node<'_>, source: &[u8]) -> Vec<String> {
     let mut targets = Vec::new();
     for index in 0..redirected.child_count() {
-        if redirected.field_name_for_child(index as u32) != Some("redirect") {
+        if redirected.field_name_for_child(index) != Some("redirect") {
             continue;
         }
-        let Some(redirect) = redirected.child(index as u32) else {
+        let Some(redirect) = redirected.child(index) else {
             continue;
         };
         if redirect.kind() != "file_redirect" {
@@ -3546,8 +3546,8 @@ fn heredoc_start_is_quoted(redirect: tree_sitter::Node<'_>, source: &[u8]) -> bo
 fn command_arguments(command: tree_sitter::Node<'_>) -> Vec<tree_sitter::Node<'_>> {
     (0..command.child_count())
         .filter_map(|index| {
-            (command.field_name_for_child(index as u32) == Some("argument"))
-                .then(|| command.child(index as u32))
+            (command.field_name_for_child(index) == Some("argument"))
+                .then(|| command.child(index))
                 .flatten()
         })
         .collect()
